pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "mtcnn>=1.0.0",
    "pillow>=11.2.1",
    "tensorflow>=2.19.0",
]
